+++ NEWS.md
@@ -1,5 +1,6 @@
 # strayr 0.1.2
 * Package renamed from `abscorr` to `strayr`.
+* `parse_income_range` added
 
 # abscorr 0.1.1
 * Functions from `strayr` package to wrangle state names + public holidays added